requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-WTF==1.2.1
argon2-cffi==23.1.0
gunicorn==21.2.0
pytest==8.0.0
